third_party/blink/web_tests/platform/generic/svg/parser/whitespace-integer-expected.txt

CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245a".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245e".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245foo".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245)90".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245em".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245ex".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245px".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245in".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245cm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245mm".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pt".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pc".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32%".
CONSOLE ERROR: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245%".
This is a testharness.js-based test.
PASS Test <integer> valid value: -47
PASS Test <integer> valid value: 0
PASS Test <integer> valid value: +32
PASS Test <integer> valid value: 1241245
PASS Test <integer> valid value: -47 
PASS Test <integer> valid value: 0 
PASS Test <integer> valid value: +32 
PASS Test <integer> valid value: 1241245 
PASS Test <integer> valid value: -47   
PASS Test <integer> valid value: 0   
PASS Test <integer> valid value: +32   
PASS Test <integer> valid value: 1241245   
PASS Test <integer> trailing garbage, value: -47a
PASS Test <integer> trailing garbage, value: 0a
PASS Test <integer> trailing garbage, value: +32a
PASS Test <integer> trailing garbage, value: 1241245a
PASS Test <integer> trailing garbage, value: -47e
PASS Test <integer> trailing garbage, value: 0e
PASS Test <integer> trailing garbage, value: +32e
PASS Test <integer> trailing garbage, value: 1241245e
PASS Test <integer> trailing garbage, value: -47foo
PASS Test <integer> trailing garbage, value: 0foo
PASS Test <integer> trailing garbage, value: +32foo
PASS Test <integer> trailing garbage, value: 1241245foo
PASS Test <integer> trailing garbage, value: -47)90
PASS Test <integer> trailing garbage, value: 0)90
PASS Test <integer> trailing garbage, value: +32)90
PASS Test <integer> trailing garbage, value: 1241245)90
PASS Test <integer> trailing garbage, value: -47em
PASS Test <integer> trailing garbage, value: 0em
PASS Test <integer> trailing garbage, value: +32em
PASS Test <integer> trailing garbage, value: 1241245em
PASS Test <integer> trailing garbage, value: -47ex
PASS Test <integer> trailing garbage, value: 0ex
PASS Test <integer> trailing garbage, value: +32ex
PASS Test <integer> trailing garbage, value: 1241245ex
PASS Test <integer> trailing garbage, value: -47px
PASS Test <integer> trailing garbage, value: 0px
PASS Test <integer> trailing garbage, value: +32px
PASS Test <integer> trailing garbage, value: 1241245px
PASS Test <integer> trailing garbage, value: -47in
PASS Test <integer> trailing garbage, value: 0in
PASS Test <integer> trailing garbage, value: +32in
PASS Test <integer> trailing garbage, value: 1241245in
PASS Test <integer> trailing garbage, value: -47cm
PASS Test <integer> trailing garbage, value: 0cm
PASS Test <integer> trailing garbage, value: +32cm
PASS Test <integer> trailing garbage, value: 1241245cm
PASS Test <integer> trailing garbage, value: -47mm
PASS Test <integer> trailing garbage, value: 0mm
PASS Test <integer> trailing garbage, value: +32mm
PASS Test <integer> trailing garbage, value: 1241245mm
PASS Test <integer> trailing garbage, value: -47pt
PASS Test <integer> trailing garbage, value: 0pt
PASS Test <integer> trailing garbage, value: +32pt
PASS Test <integer> trailing garbage, value: 1241245pt
PASS Test <integer> trailing garbage, value: -47pc
PASS Test <integer> trailing garbage, value: 0pc
PASS Test <integer> trailing garbage, value: +32pc
PASS Test <integer> trailing garbage, value: 1241245pc
PASS Test <integer> trailing garbage, value: -47%
PASS Test <integer> trailing garbage, value: 0%
PASS Test <integer> trailing garbage, value: +32%
PASS Test <integer> trailing garbage, value: 1241245%
PASS Test <integer> valid value:  -47
PASS Test <integer> valid value:  0
PASS Test <integer> valid value:  +32
PASS Test <integer> valid value:  1241245
PASS Test <integer> valid value:  -47 
PASS Test <integer> valid value:  0 
PASS Test <integer> valid value:  +32 
PASS Test <integer> valid value:  1241245 
PASS Test <integer> valid value:  -47   
PASS Test <integer> valid value:  0   
PASS Test <integer> valid value:  +32   
PASS Test <integer> valid value:  1241245   
PASS Test <integer> trailing garbage, value:  -47a
PASS Test <integer> trailing garbage, value:  0a
PASS Test <integer> trailing garbage, value:  +32a
PASS Test <integer> trailing garbage, value:  1241245a
PASS Test <integer> trailing garbage, value:  -47e
PASS Test <integer> trailing garbage, value:  0e
PASS Test <integer> trailing garbage, value:  +32e
PASS Test <integer> trailing garbage, value:  1241245e
PASS Test <integer> trailing garbage, value:  -47foo
PASS Test <integer> trailing garbage, value:  0foo
PASS Test <integer> trailing garbage, value:  +32foo
PASS Test <integer> trailing garbage, value:  1241245foo
PASS Test <integer> trailing garbage, value:  -47)90
PASS Test <integer> trailing garbage, value:  0)90
PASS Test <integer> trailing garbage, value:  +32)90
PASS Test <integer> trailing garbage, value:  1241245)90
PASS Test <integer> trailing garbage, value:  -47em
PASS Test <integer> trailing garbage, value:  0em
PASS Test <integer> trailing garbage, value:  +32em
PASS Test <integer> trailing garbage, value:  1241245em
PASS Test <integer> trailing garbage, value:  -47ex
PASS Test <integer> trailing garbage, value:  0ex
PASS Test <integer> trailing garbage, value:  +32ex
PASS Test <integer> trailing garbage, value:  1241245ex
PASS Test <integer> trailing garbage, value:  -47px
PASS Test <integer> trailing garbage, value:  0px
PASS Test <integer> trailing garbage, value:  +32px
PASS Test <integer> trailing garbage, value:  1241245px
PASS Test <integer> trailing garbage, value:  -47in
PASS Test <integer> trailing garbage, value:  0in
PASS Test <integer> trailing garbage, value:  +32in
PASS Test <integer> trailing garbage, value:  1241245in
PASS Test <integer> trailing garbage, value:  -47cm
PASS Test <integer> trailing garbage, value:  0cm
PASS Test <integer> trailing garbage, value:  +32cm
PASS Test <integer> trailing garbage, value:  1241245cm
PASS Test <integer> trailing garbage, value:  -47mm
PASS Test <integer> trailing garbage, value:  0mm
PASS Test <integer> trailing garbage, value:  +32mm
PASS Test <integer> trailing garbage, value:  1241245mm
PASS Test <integer> trailing garbage, value:  -47pt
PASS Test <integer> trailing garbage, value:  0pt
PASS Test <integer> trailing garbage, value:  +32pt
PASS Test <integer> trailing garbage, value:  1241245pt
PASS Test <integer> trailing garbage, value:  -47pc
PASS Test <integer> trailing garbage, value:  0pc
PASS Test <integer> trailing garbage, value:  +32pc
PASS Test <integer> trailing garbage, value:  1241245pc
PASS Test <integer> trailing garbage, value:  -47%
PASS Test <integer> trailing garbage, value:  0%
PASS Test <integer> trailing garbage, value:  +32%
PASS Test <integer> trailing garbage, value:  1241245%
PASS Test <integer> valid value:    -47
PASS Test <integer> valid value:    0
PASS Test <integer> valid value:    +32
PASS Test <integer> valid value:    1241245
PASS Test <integer> valid value:    -47 
PASS Test <integer> valid value:    0 
PASS Test <integer> valid value:    +32 
PASS Test <integer> valid value:    1241245 
PASS Test <integer> valid value:    -47   
PASS Test <integer> valid value:    0   
PASS Test <integer> valid value:    +32   
PASS Test <integer> valid value:    1241245   
PASS Test <integer> trailing garbage, value:    -47a
PASS Test <integer> trailing garbage, value:    0a
PASS Test <integer> trailing garbage, value:    +32a
PASS Test <integer> trailing garbage, value:    1241245a
PASS Test <integer> trailing garbage, value:    -47e
PASS Test <integer> trailing garbage, value:    0e
PASS Test <integer> trailing garbage, value:    +32e
PASS Test <integer> trailing garbage, value:    1241245e
PASS Test <integer> trailing garbage, value:    -47foo
PASS Test <integer> trailing garbage, value:    0foo
PASS Test <integer> trailing garbage, value:    +32foo
PASS Test <integer> trailing garbage, value:    1241245foo
PASS Test <integer> trailing garbage, value:    -47)90
PASS Test <integer> trailing garbage, value:    0)90
PASS Test <integer> trailing garbage, value:    +32)90
PASS Test <integer> trailing garbage, value:    1241245)90
PASS Test <integer> trailing garbage, value:    -47em
PASS Test <integer> trailing garbage, value:    0em
PASS Test <integer> trailing garbage, value:    +32em
PASS Test <integer> trailing garbage, value:    1241245em
PASS Test <integer> trailing garbage, value:    -47ex
PASS Test <integer> trailing garbage, value:    0ex
PASS Test <integer> trailing garbage, value:    +32ex
PASS Test <integer> trailing garbage, value:    1241245ex
PASS Test <integer> trailing garbage, value:    -47px
PASS Test <integer> trailing garbage, value:    0px
PASS Test <integer> trailing garbage, value:    +32px
PASS Test <integer> trailing garbage, value:    1241245px
PASS Test <integer> trailing garbage, value:    -47in
PASS Test <integer> trailing garbage, value:    0in
PASS Test <integer> trailing garbage, value:    +32in
PASS Test <integer> trailing garbage, value:    1241245in
PASS Test <integer> trailing garbage, value:    -47cm
PASS Test <integer> trailing garbage, value:    0cm
PASS Test <integer> trailing garbage, value:    +32cm
PASS Test <integer> trailing garbage, value:    1241245cm
PASS Test <integer> trailing garbage, value:    -47mm
PASS Test <integer> trailing garbage, value:    0mm
PASS Test <integer> trailing garbage, value:    +32mm
PASS Test <integer> trailing garbage, value:    1241245mm
PASS Test <integer> trailing garbage, value:    -47pt
PASS Test <integer> trailing garbage, value:    0pt
PASS Test <integer> trailing garbage, value:    +32pt
PASS Test <integer> trailing garbage, value:    1241245pt
PASS Test <integer> trailing garbage, value:    -47pc
PASS Test <integer> trailing garbage, value:    0pc
PASS Test <integer> trailing garbage, value:    +32pc
PASS Test <integer> trailing garbage, value:    1241245pc
PASS Test <integer> trailing garbage, value:    -47%
PASS Test <integer> trailing garbage, value:    0%
PASS Test <integer> trailing garbage, value:    +32%
PASS Test <integer> trailing garbage, value:    1241245%
Harness: the test ran to completion.
